Mark required fields properly
Currently, 'OPTIONAL' fields are marked as not required, but we don't
mark anything as required otherwise and you have to explicitly mark
things as required for openapi-generator to mark them as such in the
binding. This fixes that issue by marking everything as required and
then marking it as not required if it is 'OPTIONAL'.

Signed-off-by: J.R. Garcia <jrg@vmware.com>