Add default content type for request and response  (#48)
We don't specify what we produce at all ('application/json') or what we
consume at a root level. This allows us to have this specified for our
entire API.

I came across this when testing an OpenAPI mock server that takes an
OpenAPI spec and starts a mock server. When using it, it would never
correctly return anything because it didn't know what to produce.
With this commit, it will now return mock responses.

Signed-off-by: J.R. Garcia <jrg@vmware.com>