Add switch for deprecated rest APIs generation
Some of the apparent APIs contain Auto REST descriptions (obtained via
the navigation service) or @RequestMapping annotations, both of which
were later on updated with the new @VERB annotations, essentially,
making the old ones deprecated.

The following work introduces a switch - the 'mixed' parameter, as
described in the README file, which allows generation of the deprecated
APIs. Priorly, the deprecated APIs were not generated at all.

Signed-off-by: Anton Obretenov <obretenova@vmware.com>