Add switch for deprecated rest APIs generation (#57)
Add deprecate-slash-rest switch (deprecate /rest)

Some of the apparent APIs contain Auto REST descriptions (obtained via
the navigation service) or @RequestMapping annotations, both of which
were later on augmented with the new @VERB annotations (which render
on /api), making the old ones deprecated.

The following work introduces a switch - the 'deprecate-slash-rest' , as
described in the README file, which enables rendering of the /api variants 
for the API and deprecation of their /rest variants in the generated OpenAPI
definition.

Added a blacklist_utils.py utility file, which contains a blacklist
map, used for describing services that are not supposed to be
generated under /rest or /api.

Added global 'consumes' and 'produces' with values for both
application/json. The prior code, trying to place these keys per path
was removed. As a TODO is left to implement a finer-grained approach,
rather than global.

Replacement structure changed, replacements are solely identifiable by
service and operation id, pointing to method and path. Also,
code duplication reduced, formatting and name changes.

Signed-off-by: Anton Obretenov <obretenova@vmware.com>
Co-authored-by: Anton Obretenov <obretenova@vmware.com>