Remove path preprocessing, add /api on blacklisted
Path preprocessing tries to squish the paths with query params, (which
are invalid according to the spec) under a more generic path, which is
invalid behavior and may cause troubles when adding paths.

When /api and /rest are existing at the same time and -dsr is off, then
the /rest services should be shown, unless they are blacklisted.

Signed-off-by: Anton Obretenov <obretenova@vmware.com>