Add support for the @SerializationName annotation
This change adds option to replace field name in schemas with the value
of the @SerializationName. 5 tests failed and I fixed them. In one of
the tests I included @SerializationName with a value.

I have also tested against a mock service to mimic CRD.

This covers structure fields. I am unsure if we need additional fixes
for other contexts e.g. query fields. This will come in subsequent
changes.