Add structure for security definitions
Added structure containing information about the security definitions
per package,service and operation. As noted from the metadata vmodls,
the inheritence scheme in the metadata is the same as in the
specifications.

Added tests for the new structure.

TODO - fetch data from the metadata service and represent it with the
AuthenticationComponent. Finally output the AuthenticationComponent per
operation in the spec.

Signed-off-by: Anton Obretenov <obretenova@vmware.com>